backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
azure-core==1.39.0
azure-storage-blob==12.28.0
backcall==0.2.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.3.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.2
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
durationpy==0.10
ecdsa==0.19.2
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.129.0
fastjsonschema==2.21.2
filelock==3.24.2
flatbuffers==25.12.19
fsspec==2026.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipython==8.12.3
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kubernetes==35.0.0
loguru==0.7.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.1
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
networkx==3.6.1
numpy==2.4.2
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.3
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pdfminer.six==20251230
pdfplumber==0.11.9
pickleshare==0.7.5
pillow==12.1.1
pipreqs==0.5.0
platformdirs==4.9.4
posthog==5.4.0
prompt_toolkit==3.0.52
protobuf==6.33.5
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.3
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.4.0
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
python-pptx==1.0.2
PyYAML==6.0.3
pyzmq==27.1.0
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.6.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
win32_setctime==1.2.0
xlsxwriter==3.2.9
yarg==0.1.9
zipp==3.23.0
